refactor: merge MultiUserMode tests into user-mode.test.ts for coverage check
